+++ src/index.ts
@@ -1,5 +1,4 @@
 import './type-extensions';
-import chalk from 'chalk';
 import path from 'path';
 import fs from 'fs-extra';
 import murmur128 from 'murmur-128';
@@ -184,6 +183,7 @@ function createNetworkFromConfig(
     config,
     live: config.live,
     saveDeployments: config.saveDeployments,
+    zksync: config.zksync,
     tags,
     deploy: config.deploy || env.config.paths.deploy,
     companionNetworks: {},
@@ -216,6 +216,10 @@ function networkFromConfig(
       // eslint-disable-next-line @typescript-eslint/no-explicit-any
       network.verify = {etherscan: (network.config as any).etherscan};
     }
+  }
+
+  if (network.config.zksync !== undefined) {
+    network.zksync = network.config.zksync;
   }
 
   // associate tags to current network as object
@@ -888,6 +892,14 @@ task('export-artifacts')
     'solcInput',
     'if set, artifacts will have an associated solcInput files (required for old version of solidity to ensure verifiability'
   )
+  .addFlag(
+    'includingEmptyBytecode',
+    'if set, even contract without bytecode (like interfaces) will be exported'
+  )
+  .addFlag(
+    'includingNoPublicFunctions',
+    'if set, even contract without public interface (like imternal libraries) will be exported'
+  )
   .addOptionalParam(
     'exclude',
     'list of contract names separated by commas to exclude',
@@ -897,6 +909,16 @@ task('export-artifacts')
   .addOptionalParam(
     'include',
     'list of contract names separated by commas to include. If specified, only these will be considered',
+    undefined,
+    types.string
+  )
+  .addFlag(
+    'hideSources',
+    'if set, the artifacts files will not contain source code (metadata or other data exposing it) unless specified via --sources-for'
+  )
+  .addOptionalParam(
+    'sourcesFor',
+    'list of contract names separated by commas to include source (metadata,etc...) for (see --hide-sources)',
     undefined,
     types.string
   )
@@ -913,6 +935,16 @@ task('export-artifacts')
     );
     const argsExclude: string[] = args.exclude ? args.exclude.split(',') : [];
     const exclude = argsExclude.reduce(
+      (result: Record<string, boolean>, item: string) => {
+        result[item] = true;
+        return result;
+      },
+      {}
+    );
+    const argsSourcesFor: string[] = args.sourcesFor
+      ? args.sourcesFor.split(',')
+      : [];
+    const sourcesFor = argsSourcesFor.reduce(
       (result: Record<string, boolean>, item: string) => {
         result[item] = true;
         return result;
@@ -944,6 +976,21 @@ task('export-artifacts')
       const output =
         buildInfo.output.contracts[artifact.sourceName][artifactName];
 
+      if (!args.includingNoPublicFunctions) {
+        if (
+          !artifact.abi ||
+          artifact.abi.filter((v) => v.type !== 'event').length === 0
+        ) {
+          continue;
+        }
+      }
+
+      if (!args.includingEmptyBytecode) {
+        if (!artifact.bytecode || artifact.bytecode === '0x') {
+          continue;
+        }
+      }
+
       // TODO decide on ExtendedArtifact vs Artifact vs Deployment type
       // save space by not duplicating bytecodes
       if (output.evm?.bytecode?.object) {
@@ -970,9 +1017,51 @@ task('export-artifacts')
         extendedArtifact.solcInputHash = solcInputHash;
       }
 
-      fs.writeFileSync(
-        path.join(extendedArtifactFolderpath, artifactName + '.json'),
-        JSON.stringify(extendedArtifact, null, '  ')
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+      let dataToWrite: any = extendedArtifact;
+      if (args.hideSources && !sourcesFor[artifactName]) {
+        dataToWrite = {
+          // eslint-disable-next-line @typescript-eslint/no-explicit-any
+          contractName: (extendedArtifact as any).contractName,
+          // eslint-disable-next-line @typescript-eslint/no-explicit-any
+          sourceName: (extendedArtifact as any).sourceName,
+          abi: extendedArtifact.abi,
+          bytecode: extendedArtifact.bytecode,
+          deployedBytecode: extendedArtifact.deployedBytecode,
+          linkReferences: extendedArtifact.linkReferences,
+          deployedLinkReferences: extendedArtifact.deployedLinkReferences,
+          devdoc: extendedArtifact.devdoc,
+          userdoc: extendedArtifact.userdoc,
+          evm: extendedArtifact.evm
+            ? {
+                gasEstimates: extendedArtifact.evm.gasEstimates,
+                methodIdentifiers: extendedArtifact.evm.methodIdentifiers,
+              }
+            : undefined,
+        };
+      }
+
+      let filepath = path.join(
+        extendedArtifactFolderpath,
+        artifactName + '.json'
       );
+      if (dataToWrite.sourceName) {
+        if (dataToWrite.contractName) {
+          filepath = path.join(
+            extendedArtifactFolderpath,
+            dataToWrite.sourceName,
+            dataToWrite.contractName + '.json'
+          );
+        } else {
+          filepath = path.join(
+            extendedArtifactFolderpath,
+            dataToWrite.sourceName,
+            artifactName + '.json'
+          );
+        }
+      }
+
+      fs.ensureFileSync(filepath);
+      fs.writeFileSync(filepath, JSON.stringify(dataToWrite, null, '  '));
     }
   });